+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Input editor
 - Recycle threads so that savestates always see the same number of threads
 - OSD element can be positioned by the user
+- Add messages on the OSD
 
 ### Fixed
 - Don't add an extra frame when user stops the game
@@ -12,6 +13,7 @@
 - Always simulate the game window being on top-left corner so that games using global mouse coords do not desync
 - Savestates work with audio not muted
 - Change the way of hooking dlfoo functions because it broke on glibc 2.27 (now using _dl_sym to get the real dlsym)
+- Fix the time increase at frame boundary when the game had some sleep calls, so that the time remains a multiple of the framerate increment
 
 ## [1.1.1] - 2018-03-25
 ### Added